* last fixes I forgot to commit from my files :)
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@3249 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ player.cpp
@@ -2319,7 +2319,7 @@ Item* Player::createCorpse(DeathList deathList)
 
 void Player::addExhaust(uint32_t ticks, Exhaust_t type)
 {
-	if(Condition* condition = Condition::createCondition(CONDITIONID_DEFAULT, CONDITION_EXHAUST, ticks, 0, false, type))
+	if(Condition* condition = Condition::createCondition(CONDITIONID_DEFAULT, CONDITION_EXHAUST, ticks, 0, false, (uint32_t)type))
 		addCondition(condition);
 }
